requirements.txt
Pillow
requests
python-dotenv 